Revert "Injectable WebActors - Undertow"
