--- comsat-actors-undertow/build.gradle
@@ -1,6 +1,0 @@
-dependencies {
-    compile "javax.inject:javax.inject:$javaxInjectVer"
-    testCompile "com.google.inject:guice:$guiceVer"
-    testCompile "org.springframework:spring-core:$springVer"
-    testCompile "org.springframework:spring-context:$springVer"
-}
